fix[FE]: Make score update when game is over(#401)
- Move screen resize event handler to Game component
